Add dma-heap allocator
The dma-heap allocator is the successor to the ION allocator, and has
been part of the Linux kernel since version 5.6. Just like with the
ION allocator, the linux-imx kernel features an extra ioctl to retrieve
a physical address out of a DMA-BUF FD. Since ION is going to be destaged
in the future, add the dma-heap allocator to use in newer imx kernels.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>