Explicitly disallow redundant mapping attempts with different flags
It is valid to try to map an already mapped buffer. The allocators then
don't allocate, they just increase an internal refcount. The user has
to call unmap() as often as map(). However, trying to map an already mapped
buffer with different mapping flags would lead to undefined behavior and
consequently has to be defined as an error and has to be checker for.
Add checks to all allocators and extend test-alloc to verify that mapping
already mapped buffers works when using the same flags.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>